backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
duckdb>=0.10.0
pandas>=2.3.2
pandas-ta
numpy>=2.2.6
scikit-learn>=1.4.0
hmmlearn>=0.3.0
ccxt>=4.3.0
httpx>=0.27.0
openai>=1.0.0
python-telegram-bot>=20.8
